requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
alembic
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
Pillow
python-barcode
qrcode
python-dotenv
pydantic[email]
email-validator
python-jose
passlib[bcrypt]
python-multipart
psycopg2-binary
cloudinary==1.41.0
pyzbar
